api/requirements.txt
Flask==2.3.0
Flask-RESTful==0.3.10
Flask-JWT-Extended==4.4.4
stomp.py==8.1.0
requests==2.31.0
SQLAlchemy==1.4.15
psycopg2==2.9.9
Faker==24.7.1
Flask-Cors==3.0.10
marshmallow==3.18.0
marshmallow-sqlalchemy==0.25.0
Flask-SQLAlchemy==2.5.1
psycopg2==2.9.9
Werkzeug==2.3.0
google-api-core==2.18.0
google-auth==2.29.0
google-cloud-core==2.4.1
google-cloud-storage==2.16.0
